AgentCoreCodeInterpreter/requirements.txt
openai-agents
bedrock-agentcore
